change keadm cmd dir
