+++ keadm/cmd/keadm/app/cmd/util/cloudconfig.go
